TP#21800 fix broken build

+++ Code/ShipWorks.Shipping.Tests/Carriers/Asendia/AsendiaShipmentTypeTest.cs
@@ -44,7 +44,7 @@ namespace ShipWorks.Shipping.Tests.Carriers.Asendia
             testObject.ConfigurePrimaryProfile(profile);
 
             Assert.Equal(123456789, profile.Asendia.AsendiaAccountID);
-            Assert.Equal((int) AsendiaServiceType.AsendiaPriorityTracked, profile.Asendia.Service);
+            Assert.Equal(AsendiaServiceType.AsendiaPriorityTracked, profile.Asendia.Service);
             Assert.Equal((int) ShipEngineContentsType.Merchandise, profile.Asendia.Contents);
             Assert.Equal((int) ShipEngineNonDeliveryType.ReturnToSender, profile.Asendia.NonDelivery);
             Assert.False(profile.Asendia.NonMachinable);
